+++ changelog.md
@@ -2,6 +2,12 @@
 
 All notable changes to `wp-settings` will be documented in this file
 
+
+## 1.0.4.0 - 2026-02-24
+
+- Bug fix: fixed callback vulnerability 
+- Bug fix: added WARNING comment for sanitize_settings
+
 ## 1.0.3.7 - 2025-11-03
 
 - bug fix: minor sanitization bug fix